+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Bug Fixes
   sendClose success. Remove final verifyNuke as its never run in the several
   months it was in, indicating that abortConnection is 100% effective.
   Issue #161.
+* Rename `SimplePushServerProtocol` to `PushServerProtocol`. Issue #117.
 
 1.5.1 (2015-09-02)
 ==================